world-processor: change MutationChance type, add functionality to set initial epoch for world generation

+++ WorldProcessor/WorldProcessor.Application/Handlers/WorldProcessing/Dto/TileConfigDto.cs
@@ -8,7 +8,7 @@
 
         public int MutationWeight { get; set; }
 
-        public int MutationChance { get; set; }
+        public double MutationChance { get; set; }
 
         public double CrowdWeightMultiplier { get; set; }
